Updated routing, refactored fetching projects, restructured remote deployment process
The routing for different views and redirects has been rearranged and polished. The fetching of projects data in 'ProjectsList.vue' was restructured with external API client file, leading to more robust and cleaner data fetching process. The remote deployment process in the 'deploying.yml' workflow file was revamped for continuous integration and delivery, leading to a more reliable and straightforward deployment strategy.