Remove fallback.html and update manifest path
This commit deletes the fallback.html file, as it's no longer needed. The server.js file was updated to reflect this removal, and code handling fallback was simplified. The manifest.json path referenced in both server.js and the deploying.yml workflow has been updated to be under the .vite directory.