Update package versions and enhance server error message
The commit updates the version of the kneox package from "0.0.0" to "0.0.2" and adds "@types/node" dependency in both package.json and package-lock.json files. It also modifies the server error message - removing the specific error code for a more generic message. Minor changes are made in the server.js formatting and to the deployment workflow's name. An enhancement in vite.config.ts to resolve file paths has been implemented.