Refactor project fetching and update server settings
This commit modifies the ProjectsList.vue file to reactivate the fetchProjects method and disable fetchProjectsData. The server.js file is updated to improve content security policy by adding nonce, adjusting static file serving location and matching routes. It also presumes the presence of the main CSS file. The favicon and title were updated in index.html and mandatory loading of the mainCss file is enforced in index.ejs.