Refactor server routing and add nonce injection
This commit adjusts the server routing in the Express app to handle all incoming requests with a universal handler, instead of previously targeting only the root and index.html paths. Moreover, to improve security, a nonce is now injected into script and style tags read directly from an HTML file. This replaces the previously used EJS for rendering, which depended on a Manifest file.