Refactor codebase and update functionality
This commit primarily removes redundant or outdated codes from multiple files and updates the functionality of organizations and employees list. Numerous files have been updated to improve the fetch and save functionalities for both organizations and employees and to enhance the structure of data. Various API client files have been deleted as part of a wider effort to streamline the codebase. The 'projectStore.ts' has been heavily modified to improve the fetch and save methods, and to adjust how data is returned for use in other parts of the application.